+++ channel/http_callback_channel.go
@@ -4,30 +4,18 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"math/rand"
 	"net/http"
-	"strconv"
 	"sync"
 	"time"
 
-	"github.com/breez/notify/notification"
-	"github.com/gin-gonic/gin"
+	"github.com/breez/notify/notify"
 	"github.com/google/martian/v3/log"
 )
 
 const (
 	callbackTimeout = 60 * time.Second
 )
-
-type ChannelNotifier interface {
-	Notify(c context.Context, request *notification.Notification) error
-}
-
-type WebhookChannel interface {
-	AddRouter(r *gin.RouterGroup)
-	Notify(c context.Context, n ChannelNotifier, request *notification.Notification) (string, error)
-}
 
 type PendingRequest struct {
 	id     uint64
@@ -38,7 +26,6 @@ type HttpCallbackChannel struct {
 	sync.Mutex
 	httpClient      *http.Client
 	callbackBaseURL string
-	basePath        string
 	random          *rand.Rand
 	pendingRequests map[uint64]*PendingRequest
 }
@@ -47,7 +34,6 @@ func NewHttpCallbackChannel(callbackBaseURL string) *HttpCallbackChannel {
 	channel := &HttpCallbackChannel{
 		httpClient:      http.DefaultClient,
 		callbackBaseURL: callbackBaseURL,
-		basePath:        "",
 		random:          rand.New(rand.NewSource(time.Now().UnixNano())),
 		pendingRequests: make(map[uint64]*PendingRequest),
 	}
@@ -55,14 +41,9 @@ func NewHttpCallbackChannel(callbackBaseURL string) *HttpCallbackChannel {
 	return channel
 }
 
-func (p *HttpCallbackChannel) AddRouter(r *gin.RouterGroup) {
-	p.basePath = r.BasePath()
-	p.addRouter(r)
-}
-
-func (p *HttpCallbackChannel) Notify(c context.Context, n ChannelNotifier, request *notification.Notification) (string, error) {
+func (p *HttpCallbackChannel) Notify(c context.Context, notifier *notify.Notifier, basePath string, request *notify.Notification) (string, error) {
 	reqID := p.random.Uint64()
-	callbackURL := fmt.Sprintf("%s/%s/response/%d", p.callbackBaseURL, p.basePath, reqID)
+	callbackURL := fmt.Sprintf("%s/%s/response/%d", p.callbackBaseURL, basePath, reqID)
 	request.Data["reply_url"] = callbackURL
 
 	pendingRequest := &PendingRequest{
@@ -85,7 +66,7 @@ func (p *HttpCallbackChannel) Notify(c context.Context, n ChannelNotifier, reque
 
 	log.Debugf("waiting for response: %v", callbackURL)
 
-	if err := n.Notify(c, request); err != nil {
+	if err := notifier.Notify(c, request); err != nil {
 		log.Debugf("failed to notify, request: %v, error: %v", request, err)
 		return "", err
 	}
@@ -100,32 +81,7 @@ func (p *HttpCallbackChannel) Notify(c context.Context, n ChannelNotifier, reque
 	}
 }
 
-func (p *HttpCallbackChannel) addRouter(r *gin.RouterGroup) {
-	r.POST("/response/:responseId", func(c *gin.Context) {
-		responseId := c.Param("responseId")
-
-		reqId, err := strconv.ParseUint(responseId, 10, 64)
-		if err != nil {
-			c.AbortWithError(http.StatusBadRequest, errors.New("invalid response"))
-			return
-		}
-
-		all, err := io.ReadAll(c.Request.Body)
-		if err != nil {
-			c.AbortWithError(http.StatusInternalServerError, errors.New("internal error"))
-			return
-		}
-
-		if err := p.onResponse(reqId, string(all)); err != nil {
-			c.AbortWithError(http.StatusInternalServerError, err)
-			return
-		}
-
-		c.Status(http.StatusOK)
-	})
-}
-
-func (p *HttpCallbackChannel) onResponse(reqID uint64, payload string) error {
+func (p *HttpCallbackChannel) OnResponse(reqID uint64, payload string) error {
 	p.Lock()
 	defer p.Unlock()
 	pendingRequest, ok := p.pendingRequests[reqID]
