--- notification/notification.go
@@ -1,10 +1,0 @@
-package notification
-
-type Notification struct {
-	Template         string
-	DisplayMessage   string
-	Type             string
-	TargetIdentifier string
-	AppData          *string
-	Data             map[string]interface{}
-}
