+++ notify/notify_test.go
@@ -5,20 +5,19 @@ import (
 	"testing"
 
 	"github.com/breez/notify/config"
-	"github.com/breez/notify/notification"
 	"gotest.tools/v3/assert"
 )
 
 type TestService struct {
-	sentQueue chan *notification.Notification
+	sentQueue chan *Notification
 }
 
 func newTestService() *TestService {
-	queue := make(chan *notification.Notification, 10)
+	queue := make(chan *Notification, 10)
 	return &TestService{sentQueue: queue}
 }
 
-func (t *TestService) Send(c context.Context, notification *notification.Notification) error {
+func (t *TestService) Send(c context.Context, notification *Notification) error {
 	t.sentQueue <- notification
 	return nil
 }
@@ -27,14 +26,14 @@ func TestNotify(t *testing.T) {
 	service := newTestService()
 	config := &config.Config{WorkersNum: 2}
 	notifier := NewNotifier(config, map[string]Service{"test": service})
-	n := notification.Notification{
+	n := Notification{
 		Template:         "t1",
 		Type:             "test",
 		TargetIdentifier: "token1",
 	}
 	notifier.Notify(context.Background(), &n)
 
-	var notifications []notification.Notification
+	var notifications []Notification
 	res := <-service.sentQueue
 	notifications = append(notifications, *res)
 	assert.Assert(t, len(notifications) == 1)
