+++ notify/services/fcm.go
@@ -6,14 +6,14 @@ import (
 	"fmt"
 
 	"firebase.google.com/go/messaging"
-	"github.com/breez/notify/notification"
+	"github.com/breez/notify/notify"
 )
 
 var (
 	ErrUnrecognizedTemplate = errors.New("unrecognized template")
 )
 
-type FCMMessageBuilder func(req *notification.Notification) (*messaging.Message, error)
+type FCMMessageBuilder func(req *notify.Notification) (*messaging.Message, error)
 type FCM struct {
 	messageBuilder FCMMessageBuilder
 	client         *messaging.Client
@@ -23,7 +23,7 @@ func NewFCM(messageBuilder FCMMessageBuilder, client *messaging.Client) *FCM {
 	return &FCM{messageBuilder: messageBuilder, client: client}
 }
 
-func (f *FCM) Send(context context.Context, req *notification.Notification) error {
+func (f *FCM) Send(context context.Context, req *notify.Notification) error {
 	pushNotification, err := f.messageBuilder(req)
 	if err != nil {
 		return fmt.Errorf("failed to create message %v", err)
